Retinal Blood Vessel Extraction
Applying an operation to a picture and extracting or changing the data for a particular application is known as image processing. Since many businesses, including those in the medical, agricultural, and factory automation sectors, already employ image processing to rectify damaged images, it can be approved and used in new industries.
Retinal blood vessels have a significant role in the diagnosis and treatment of various retinal diseases such as diabetic retinopathy, glaucoma, arteriosclerosis, and hypertension. For this reason, retinal vasculature extraction is important to help specialists for the diagnosis and treatment of systematic diseases. In this project, a novel approach is developed to extract retinal blood vessel network. Our method comprises four stages:
•	Clear the image
•	Remove small spots
•	Sharpen and prominent the vessel
•	Segment using Kirsch’s Templates